Fix Kategori Produk

+++ app/Http/Controllers/ProductCategoryController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 // Import Class ProductCategoryRequest
 use App\Http\Requests\ProductCategoryRequest;
-
 // Import Model ProductCategory
 use App\Models\ProductCategory;
 
@@ -14,7 +13,6 @@ class ProductCategoryController extends Controller
     public function index()
     {
         $productCategories = ProductCategory::with('user')->latest()->get();
-
         return view('dashboard.product_category', compact('productCategories'));
     }
 
@@ -28,9 +26,7 @@ class ProductCategoryController extends Controller
     public function store(ProductCategoryRequest $request)
     {
         $data = $request->all();
-
         $request->user()->product_categories()->create($data);
-
         return redirect('/product-category')->with('msg', 'Data Produk Kategoti Berhasil ditambahkan');
     }
 
@@ -44,11 +40,7 @@ class ProductCategoryController extends Controller
     public function edit($id)
     {
         $productCategory = ProductCategory::findOrFail($id);
-
-        if (!$productCategory->isOwner()) {
-            return redirect('/product-category')->with('msg', 'Anda tidak memiliki akses');
-        }
-
+        $this->authorize('isOwner', $productCategory);
         return view('dashboard_edit.product_category_edit', compact('productCategory'));
     }
 
@@ -56,17 +48,16 @@ class ProductCategoryController extends Controller
     public function update(ProductCategoryRequest $request, $id)
     {
         $data = $request->all();
-
         ProductCategory::findOrFail($id)->update($data);
-
         return redirect('/product-category')->with('msg', 'Data Produk Kategoti Berhasil diedit');
     }
 
     // DELETE
     public function destroy($id)
     {
+        $productCategory = ProductCategory::findOrFail($id);
+        $this->authorize('isOwner', $productCategory);
         ProductCategory::destroy($id);
-
         return redirect('/product-category')->with('msg', 'Data Produk Kategoti Berhasil dihapus');
     }
 }
